+++ src/main/java/at/srsyntax/farmingworld/FarmingWorldPlugin.java
@@ -3,7 +3,6 @@ package at.srsyntax.farmingworld;
 import at.srsyntax.farmingworld.api.API;
 import at.srsyntax.farmingworld.api.DisplayPosition;
 import at.srsyntax.farmingworld.api.DisplayType;
-import at.srsyntax.farmingworld.api.FarmingWorld;
 import at.srsyntax.farmingworld.command.FarmingCommand;
 import at.srsyntax.farmingworld.command.FarmingWorldInfoCommand;
 import at.srsyntax.farmingworld.command.FarmingWorldResetCommand;
@@ -15,13 +14,10 @@ import at.srsyntax.farmingworld.config.PluginConfig;
 import at.srsyntax.farmingworld.listener.ActionBarListeners;
 import at.srsyntax.farmingworld.listener.BossBarListeners;
 import at.srsyntax.farmingworld.listener.ConfirmListener;
-import at.srsyntax.farmingworld.runnable.date.DateCheckRunnable;
-import at.srsyntax.farmingworld.runnable.date.DateDisplayRunnable;
-import at.srsyntax.farmingworld.runnable.date.DateRunnable;
-import at.srsyntax.farmingworld.runnable.remaining.RemainingRunnable;
+import at.srsyntax.farmingworld.runnable.RunnableManager;
+import at.srsyntax.farmingworld.util.FarmingWorldLoader;
 import at.srsyntax.farmingworld.util.ResetData;
 import lombok.Getter;
-import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.boss.BarColor;
@@ -32,7 +28,6 @@ import org.bukkit.plugin.java.JavaPlugin;
 import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeUnit;
 
 /*
  * MIT License
@@ -64,8 +59,8 @@ public class FarmingWorldPlugin extends JavaPlugin {
   @Getter private static API api;
 
   @Getter private PluginConfig pluginConfig;
-  private Timer timer;
   @Getter private final Map<CommandSender, ResetData> needConfirm = new ConcurrentHashMap<>();
+  private RunnableManager runnableManager;
 
   @Override
   public void onEnable() {
@@ -75,9 +70,10 @@ public class FarmingWorldPlugin extends JavaPlugin {
       new Metrics(this, BSTATS_ID);
       loadFarmingWorlds();
 
-      startScheduler();
+      this.runnableManager = new RunnableManager(api, this);
+      this.runnableManager.startScheduler();
       registerListeners();
-      registerCommands();
+      registerCommands(pluginConfig.getMessage());
 
     } catch (Exception exception) {
       getLogger().severe("Plugin could not be loaded successfully!");
@@ -87,13 +83,10 @@ public class FarmingWorldPlugin extends JavaPlugin {
 
   @Override
   public void onDisable() {
-    if (this.timer != null)
-      this.timer.purge();
+    this.runnableManager.purge();
   }
 
-  private void registerCommands() {
-    final MessageConfig messageConfig = this.pluginConfig.getMessage();
-
+  private void registerCommands(MessageConfig messageConfig) {
     getCommand("farming").setExecutor(new FarmingCommand(api, this));
     getCommand("farmingworldinfo").setExecutor(new FarmingWorldInfoCommand(api, messageConfig));
     getCommand("farmingworldreset").setExecutor(new FarmingWorldResetCommand(api, this, messageConfig));
@@ -109,70 +102,10 @@ public class FarmingWorldPlugin extends JavaPlugin {
     pluginManager.registerEvents(new ConfirmListener(this), this);
   }
 
-  private void startScheduler() {
-    if (pluginConfig.getDisplayType() == DisplayType.REMAINING)
-      startRemainingScheduler();
-    else
-      startDateScheduler();
-  }
-
-  private void startRemainingScheduler() {
-    final Runnable runnable = new RemainingRunnable(this);
-    getServer().getScheduler().runTaskTimerAsynchronously(this, runnable, 90L, 1200L);
-  }
-
-  private void startDateScheduler() {
-    this.timer = new Timer();
-    api.getFarmingWorlds().forEach(farmingWorld -> {
-      final Date checkDate = new Date(farmingWorld.getReset() - TimeUnit.SECONDS.toMillis(5));
-      this.timer.schedule(new DateCheckRunnable(api, farmingWorld), checkDate);
-      this.timer.schedule(new DateRunnable(api, farmingWorld), new Date(farmingWorld.getReset()));
-
-      farmingWorld.updateDisplay();
-    });
-  }
 
   private void loadFarmingWorlds() {
-    this.pluginConfig.getFarmingWorlds().forEach(farmingWorld -> {
-      farmingWorld.setPlugin(this);
-      checkBorder(farmingWorld);
-      checkCurrentWorld(farmingWorld);
-      checkNextWorld(farmingWorld);
-    });
-
-    if (this.pluginConfig.getDateRefresh() > 0) {
-      final long time = this.pluginConfig.getDateRefresh() * 20L;
-      Bukkit.getScheduler().runTaskTimerAsynchronously(this, new DateDisplayRunnable(api), time, time);
-    }
-  }
-
-  private void checkBorder(FarmingWorld farmingWorld) {
-    if (farmingWorld.getBorderSize() < 10) return;
-    if (farmingWorld.getBorderSize() >= farmingWorld.getRtpArenaSize()) return;
-    final String prefix = farmingWorld.getName() + ": ";
-    final int newRTPArenaSize = (int) (farmingWorld.getBorderSize() - 2);
-    getLogger().severe(prefix + "The RTP arena size must be smaller than the world border.");
-    getLogger().severe(prefix + "The RTP arena size is therefore changed from " + farmingWorld.getRtpArenaSize() + " to " + newRTPArenaSize + ".");
-    ((FarmingWorldConfig) farmingWorld).setRtpArenaSize(newRTPArenaSize);
-  }
-
-  private void checkCurrentWorld(FarmingWorldConfig farmingWorld) {
-    if (farmingWorld.getCurrentWorldName() == null) {
-      farmingWorld.newWorld(api.generateFarmingWorld(farmingWorld));
-    } else {
-      api.loadFarmingWorld(farmingWorld.getCurrentWorldName(), farmingWorld.getEnvironment());
-    }
-  }
-
-  private void checkNextWorld(FarmingWorldConfig farmingWorld) {
-    if (farmingWorld.getNextWorldName() == null) {
-      final long remaining = farmingWorld.getReset() - System.currentTimeMillis();
-      if (farmingWorld.needReset() || (remaining <= TimeUnit.MINUTES.toMillis(5) && remaining > 0)) {
-        farmingWorld.setNextWorld(api.generateFarmingWorld(farmingWorld));
-      }
-    } else {
-      api.loadFarmingWorld(farmingWorld.getNextWorldName(), farmingWorld.getEnvironment());
-    }
+    final FarmingWorldLoader loader = new FarmingWorldLoader(getLogger(), api, this);
+    this.pluginConfig.getFarmingWorlds().forEach(loader::load);
   }
 
   private PluginConfig loadConfig() throws IOException {
